update lpn to use uint64_t

+++ sims/lpn/lpn_common/lpn_sim.hh
@@ -1,13 +1,14 @@
 #ifndef __LPN_SIM__
 #define __LPN_SIM__
+#include <bits/stdint-uintn.h>
 #include "place_transition.hh"
 #define LOOP_TS(func, t_size) for(int i=0;i < t_size; i++){ \
         Transition* t = t_list[i]; \
         func; \
     } 
 
-int NextCommitTime(Transition* t_list[], int size);
+uint64_t NextCommitTime(Transition* t_list[], int size);
 
-int CommitAtTime(Transition* t_list[], int size, int time);
+int CommitAtTime(Transition* t_list[], int size, uint64_t time);
 
 #endif
